Refactor raw types
