+++ doc/changelog/1.2.0.md
@@ -27,6 +27,10 @@
     * `seamless`
     * `selected`
 
+* Renamed:
+  * `.raw.ReactChildren` => `.raw.React.Children`
+  * `.raw.ReactComponentEs6` => `.raw.React.Component`
+
 * Removed
   * `ScalaComponent.Builder#buildWithReactCreateClass` -  `React.createClass` was removed in React 16
